// src/app/models/playerfull.ts
export class PlayerFull {
    //player info
    filtro1: string;
    filtro2: string;
    idplayer: number;
    idTeam: number;
    idOwner: number;
    firstName: string;
    lastName: string;
    idnationality: number;
    namenationality: string;
    age: number;
    height: number;
    dmi: number;
    jersey: number;
    salary: number;
    bestPosition: string;
    bestPositionNew: string;
    seasonDrafted: number;
    leagueDrafted: number;
    teamDrafted: number;
    draftPick: number;
    forSale: number;
    sumaInterior: number;
    sumaExterior: number;
    sumaTotal: number;
    //skill info
    gameShape: number;
    potential: number;
    jumpShot: number;
    nrange: number;
    outsideDef: number;
    handling: number;
    driving: number;
    passing: number;
    insideShot: number;
    insideDef: number;
    rebound: number;
    block: number;
    stamina: number;
    freeThrow: number;
    experience: number;    
    flag_gameShape: number;
    flag_potential: number;
    flag_jumpShot: number;
    flag_nrange: number;
    flag_outsideDef: number;
    flag_handling: number;
    flag_driving: number;
    flag_passing: number;
    flag_insideShot: number;
    flag_insideDef: number;
    flag_rebound: number;
    flag_block: number;
    flag_stamina: number;
    flag_freeThrow: number;
    flag_experience: number;
    h_gameShape: number;
    h_potential: number;
    h_jumpShot: number;
    h_nrange: number;
    h_outsideDef: number;
    h_handling: number;
    h_driving: number;
    h_passing: number;
    h_insideShot: number;
    h_insideDef: number;
    h_rebound: number;
    h_block: number;
    h_stamina: number;
    h_freeThrow: number;
    h_experience: number;
    //stat info
    games: number;	
    mpg: number;	
    fgPerc: number;	
    tpPerc: number;	
    ftPerc: number;	
    orpg: number;	
    rpg: number;	
    apg: number;	
    topg: number;	
    spg: number;	
    bpg: number;	
    ppg: number;	
    fpg: number;	
    rating: number;	
    flag_games: number;	
    flag_mpg: number;	
    flag_fgPerc: number;	
    flag_tpPerc: number;	
    flag_ftPerc: number;	
    flag_orpg: number;	
    flag_rpg: number;	
    flag_apg: number;	
    flag_topg: number;	
    flag_spg: number;	
    flag_bpg: number;	
    flag_ppg: number;	
    flag_fpg: number;	
    flag_rating: number;
    teamName: string;	
    check1: number;
    check2: number;
    owner: string;
    dupdate: any;
    flagFecha3Dias: number;
}